🐛 Fix types issues in editor
At the moment when why try to run test in editor we get this error:
```
There are types at '@reedsy/express-ws/index.d.ts', but this result could not be resolved when respecting package.json "exports".
```

This was not a problem before as other apps does not use `@reedsy/express-ws` directly.